+++ src/lizaalert/courses/mixins.py
@@ -1,8 +1,6 @@
 from django.core.validators import MinValueValidator
 from django.db import models
-from django.db.models import F, Max
-
-from lizaalert.courses.exceptions import WrongMethodException
+from django.db.models import Max
 
 
 class TimeStampedModel(models.Model):
@@ -81,43 +79,69 @@ def order_number_mixin(step, parent_field):
             if not self.id:
                 max_order_number = queryset.aggregate(Max("order_number")).get("order_number__max")
                 self.order_number = (max_order_number or 0) + order_factor
+                return self.order_number
+
+            # получаем старый порядковый номер
+            old_order_number = type(self).objects.filter(id=self.id).values("order_number").first()
+            # округляем новый порядковый номер до шага очередности
+            if self.order_number % order_factor != 0:
+                rounded_new_order_number = (self.order_number // order_factor + 1) * order_factor
             else:
-                # получаем старый порядковый номер
-                old_order_number = type(self).objects.get(id=self.id)
-                # округляем новый порядковый номер до шага очередности
-                if self.order_number % 10 != 0:
-                    rounded_new_order_number = (self.order_number // order_factor + 1) * order_factor
-                else:
-                    rounded_new_order_number = self.order_number
-                if old_order_number.order_number != self.order_number:
-                    objects = queryset.exclude(id=self.id).order_by("order_number")
-                    objects_to_update = []
-                    modifier_flag = False
-                    for position, object in enumerate(objects):
-                        if modifier_flag:
-                            object.order_number = (position + 2) * order_factor
-                            objects_to_update.append(object)
+                rounded_new_order_number = self.order_number
+            if old_order_number["order_number"] != self.order_number:
+                objects = queryset.exclude(id=self.id).order_by("order_number")
+                objects_to_update = []
+                modifier_flag = False
+                # for position, object in enumerate(objects):
+                #     if modifier_flag:
+                #         object.order_number = (position + 2) * order_factor
+                #         objects_to_update.append(object)
 
-                        # переносим все последующие уроки на один шаг вперед
-                        elif object.order_number == rounded_new_order_number:
-                            object.order_number = (position + 2) * order_factor
-                            modifier_flag = True
-                            objects_to_update.append(object)
+                #     # переносим все последующие уроки на один шаг вперед
+                #     elif object.order_number == rounded_new_order_number:
+                #         object.order_number = (position + 2) * order_factor
+                #         modifier_flag = True
+                #         objects_to_update.append(object)
 
-                        # если есть окно, заполняем его уроками
-                        elif (position + 1) * order_factor != object.order_number:
-                            object.order_number = (position + 1) * order_factor
-                            objects_to_update.append(object)
+                #     # если есть окно, заполняем его уроками
+                #     elif (position + 1) * order_factor != object.order_number:
+                #         object.order_number = (position + 1) * order_factor
+                #         objects_to_update.append(object)
+                position = order_factor
+                anchor = 0
+                while objects:
+                    obj = objects[anchor]
                     if modifier_flag:
-                        self.order_number = rounded_new_order_number
+                        obj.order_number = position
+                        objects_to_update.append(obj)
+
+                    # переносим все последующие уроки на один шаг вперед
+                    elif obj.order_number == rounded_new_order_number:
+                        position += order_factor
+                        obj.order_number = position
+                        modifier_flag = True
+                        objects_to_update.append(obj)
+
+                    # если есть окно, заполняем его уроками
+                    elif position != obj.order_number:
+                        obj.order_number = position
+                        objects_to_update.append(obj)
+
+                    position += order_factor
+                    anchor += 1
+                    if anchor >= len(objects):
+                        break
+                if modifier_flag:
+                    self.order_number = rounded_new_order_number
+                else:
+                    # если менялся только порядок последнего урока, то назначаем номер от последнего-1 урока
+                    if objects_to_update:
+                        max_order_number = max(obj.order_number for obj in objects_to_update)
                     else:
-                        # если менялся только порядок последнего урока, то назначаем номер от последнего-1 урока
-                        if objects_to_update:
-                            max_order_number = max(obj.order_number for obj in objects_to_update)
-                        else:
-                            max_order_number = max(obj.order_number for obj in objects)
-                        self.order_number = max_order_number + order_factor
-                    queryset.model.objects.bulk_update(objects_to_update, ["order_number"])
+                        max_order_number = max(obj.order_number for obj in objects)
+                    self.order_number = max_order_number + order_factor
+                queryset.model.objects.bulk_update(objects_to_update, ["order_number"])
+            return self.order_number
 
         def save(self, *args, **kwargs):
             """Change ordering method."""
@@ -127,33 +151,5 @@ def order_number_mixin(step, parent_field):
             # else:
             #     self.reset_ordering(self.order_queryset, step)
             super().save(*args, **kwargs)
-
-        @property
-        def ordered(self):
-            """Вернуть очередность всех уроков курса с полем ordering."""
-            if parent_field == "chapter":
-                cls = type(self)
-                return (
-                    cls.objects.filter(chapter__course=self.chapter.course)
-                    .annotate(ordering=F("chapter__order_number") + F("order_number"))
-                    .order_by("ordering")
-                )
-            raise WrongMethodException
-
-        @property
-        def next_lesson(self):
-            """Вернуть следующий по очереди урок."""
-            if parent_field == "chapter":
-                ordered_lessons = self.ordered
-                return ordered_lessons.filter(ordering__gt=self.ordering).order_by("ordering").values("id")[:1]
-            raise WrongMethodException
-
-        @property
-        def prev_lesson(self):
-            """Вернуть предыдущий по очереди урок."""
-            if parent_field == "chapter":
-                ordered_lessons = self.ordered
-                return ordered_lessons.filter(ordering__lt=self.ordering).order_by("-ordering").values("id")[:1]
-            raise WrongMethodException
 
     return SaveOrderingMixin
